bind/requirements.txt
# integration pip requirements
dnspython>=1.15.0
